ConversionUtilsTest: rewrite testBadObjectElements
This test is no longer applicable, at least with the current conversion
logic that now handles arrays and collections as of #449 et al.

There is a substantial discrepancy right now between what
DefaultConverter reports as convertible via canConvert versus
what it actually *does* support converting via the convert methods.
But that's an issue for another day!