Stop assigning ConversionUtils delegates
This logic is inherently unsound, because it pollutes static state.
It changes the behavior of the ConversionUtils methods depending on
whether a SciJava Context with a ConvertService has ever been
instantiated, and depending on the available Converter plugins.

With this logic removed, all tests now pass, and pass more consistently,
wheras before, the ConversionUtilsTest would have all passing tests when
run standalone, and failing test(s) when run as part of the suite.

I'm aware that this change may break some things downstream. But because
no unit tests exist here in scijava-common to catch those theoretical
cases, we will just have to give it a try by making a release, and then
address any issues arising afterward.