SystemPropertyArgument: allow any string for keys
The previous regex was not allowing colons, which Java definitely
allows. But after briefly trying to research what's allowed and
what's not, I decided it doesn't really matter, and we should just
try setting whatever the user is trying to set, and if it fails with
an exception on the Java side, so be it -- but that's no reason not
to eagerly reject the expression in this ConsoleArgument plugin.