DefaultThreadService: let running tasks finish
We wait up to 5 seconds for them to terminate.

I don't really have a good idea whether this is a good amount of time,
but it should be more than enough for common scenarios.

At some point, the ThreadService interface could gain methods to
override this value, but for now, we hardcode it.