ScriptREPL: generalize input and output streams
The use of InputStream and OutputStream was natural, but those are heavy
abstract classes, not lightweight interfaces. It is more flexible to
allow specifying a Supplier from which to draw the REPL input, and/or a
Consumer to receive the REPL output. This will be useful e.g. from
Python via JPype, which allows to implement Java interfaces in Python
backed by proxies. But you cannot proxy an abstract class like
InputStream or OutputStream. Nonetheless, the ability to simply use
InputStream/OutputStream like System.in+System.out is also nice,
so we keep those method signatures as well.