UIService: run UI code on the EDT as needed
With 96827bf5ea2400b605cacad5e6dbf9984441c06f, the showUI methods now
run on the EDT if the active UI requires it. But the other UI
mutators (show, showDialog, chooseFile, chooseFiles, showContextMenu)
were still being run naively on the curent thread. This commit fixes
the behavior to be consistent with showUI.