De-duplicate widget inputs by String comparison
When populating potential items for an input widget, we now prefer
existing objects of a given input type to potentially convertiable
types.

Further, for convertibles we now avoid considering them if they share a
toString with any other potential input. Candidates are prioritized in
order returned by ConvertService.getCompatibleInputs.

This mitigates the potential for duplicate entries in input harvesters
when multiple input instances are convertible to the same effective
output instance.