Try to be thread-safe when showing UI elements
Instead of being thread-safe in the DefaultUIService, we move the thread
safety into the AbstractUserInterface itself, since each user interface
is responsible for deciding how to behave. This retains the behavior of
showUI being thread-safe, but also makes the other ui.show methods
thread-safe as well. So now DefaultUIService doesn't need to care.