Support similar ScriptLanguages
If we have multiple languages that support the same extension/name we
now use the highest priority ScriptLanguage that isActive (a new helper
method for ScriptLanguages) instead of overwriting the language for that
extension/name.

This allows multiple language implementations to co-exist and take
priority based on the execution context (e.g. Jython works in Java-based
environments, PyImageJ works in python mode apps).