Add KeyCode.get(char) method
Previously, passing a char to KeyCode would call KeyCode.get(int), which
does not do what one might expect. For example, calling KeyCode.get('a')
would convert the 'a' character to integer/ASCII code 97, which
corresponds to KeyCode.NUMPAD_0!

This commit introduces a function that behaves as one would expect
KeyCode.get(char) to behave: mapping the actual character value to the
key code constant associated with it. So now, KeyCode.get('a') returns
KeyCode.A as one might expect.

Unfortunately, not all codes have clean character values, and not all
characters map unambiguously to key codes; e.g.:

* Both 'a' and 'A' map to KeyCode.A
* '0' maps to KeyCode.NUM0, not KeyCode.NUMPAD_0
* '/' maps to KeyCode.SLASH, not KeyCode.NUMPAD_SLASH

But it's still a big step forward in abiding by PoLA.